requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.12.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
gunicorn
psycopg2-binary==2.8.6
django-environ

